+++ base/Color.h
@@ -18,13 +18,13 @@ public:
   const static Color kBlack;
 
   Color();
-  Color(float r, float g, float b);
-  Color Multiply(float a) const;
+  Color(double r, double g, double b);
+  Color Multiply(double a) const;
   Color Modulate(const Color &c) const;
   Color Add(const Color &c) const;
   Color Sub(const Color &c) const;
 
-  float red, green, blue;
+  double red, green, blue;
 };
 
 #endif /* COLOR_H_ */
