+++ materials/Material.h
@@ -14,13 +14,13 @@ class Material {
 public:
   Material();
   Material(Material & m);
-  Material(float _diffusion, float _reflection, float _refraction,
+  Material(double _diffusion, double _reflection, double _refraction,
            Color _emittance, Color _color);
   virtual ~Material();
 
-  float diffusion;
-  float reflection;
-  float refraction;
+  double diffusion;
+  double reflection;
+  double refraction;
   Color emittance;
   Color color;
